refcotr :: 표 부분 컴포넌트화 시키고 임포트
